Move scales to context to avoid circular deps
