+++ packages/context/package.json
@@ -36,7 +36,7 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "dependencies": {
-    "@times-components/styleguide": "3.0.1"
+    "@times-components/styleguide": "3.1.1"
   },
   "devDependencies": {
     "@times-components/eslint-config-thetimes": "0.7.1",
