+++ packages/context/src/context.js
@@ -2,5 +2,4 @@ import { createContext } from "react";
 import defaults from "./defaults";
 
 export default createContext(defaults);
-export { default as scales } from "./defaults/scales";
 export { defaults };
