+++ packages/pages/package.json
@@ -54,9 +54,9 @@
   "dependencies": {
     "@times-components/article": "4.1.3",
     "@times-components/author-profile": "4.0.6",
-    "@times-components/context": "0.2.3",
+    "@times-components/context": "0.3.1",
     "@times-components/provider": "1.0.3",
-    "@times-components/styleguide": "3.0.1",
+    "@times-components/styleguide": "3.1.1",
     "@times-components/topic": "3.0.6",
     "@times-components/utils": "3.0.1",
     "apollo-cache-inmemory": "1.1.10",
