+++ packages/styleguide/__tests__/android/styleguide.android.test.js
@@ -1,6 +1,5 @@
-import { scales } from "@times-components/context";
 import { iterator } from "@times-components/test-utils";
-import styleguide from "../../src/styleguide";
+import styleguide, { scales } from "../../src/styleguide";
 import shared from "../shared";
 import sharedNative from "../shared-native";
