+++ packages/styleguide/package.json
@@ -61,7 +61,6 @@
     "webpack-cli": "2.1.4"
   },
   "dependencies": {
-    "@times-components/context": "0.3.1",
     "@times-components/responsive-styles": "3.0.2",
     "prop-types": "15.6.2"
   },
